Инпут категорий вынесен в отдельный компонент
