+++ src/components/TransactionsExpForm/TransactionsExpForm.module.scss
@@ -103,69 +103,3 @@
     border: 2px solid #f5f6fb;
   }
 }
-
-.form_category_list {
-  position: absolute;
-  left: 0px;
-  top: 100%;
-  color: #c7ccdc;
-  box-shadow: 0px 3px 4px rgba(170, 178, 197, 0.4);
-  background-color: white;
-  border: none;
-}
-.form_input_category_parent {
-  position: relative;
-}
-
-.category_label {
-  padding: 9px 20px;
-  font-family: Roboto;
-  font-style: normal;
-  font-weight: normal;
-  font-size: 12px;
-  line-height: 14px;
-  color: #c7ccdc;
-  height: 32px;
-  width: 188px;
-  display: inline-block;
-  cursor: pointer;
-
-  &:hover,
-  &:focus {
-    background-color: #f5f6fb;
-    color: black;
-  }
-}
-.arrow {
-  position: absolute;
-  right: 18px;
-  top: 18px;
-  width: 12px;
-  height: 8px;
-  color: #52555f;
-}
-.arrow_down {
-  @extend .arrow;
-  background-image: url("../../images/icons/arrow-down.svg");
-}
-.arrow_up {
-  @extend .arrow;
-  background-image: url("../../images/icons/arrow-up.svg");
-}
-.form_input_category {
-  position: relative;
-  width: 188px;
-  height: 44px;
-  border: 2px solid #f5f6fb;
-  border-top: 2px solid #f5f6fb;
-  border-right: none;
-  border-bottom: 2px solid #f5f6fb;
-  border-left: none;
-  box-sizing: border-box;
-  outline: none;
-  padding-left: 21px;
-
-  &:focus + .arrow {
-    transform: rotate(-180deg);
-  }
-}
